move handler to impl subpackage, add generic service operator
