+++ webapp/src/main/resources/components/skeleton.xml
@@ -7,9 +7,14 @@
                            http://www.springframework.org/schema/util
                            http://www.springframework.org/schema/util/spring-util-4.1.xsd">
 
+    <bean class="org.n52.iceland.service.operator.GenericServiceOperator">
+        <constructor-arg name="service" value="#{T(org.n52.iceland.skeleton.SkeletonConstants).SERVICE}" />
+        <constructor-arg name="version" value="#{T(org.n52.iceland.skeleton.SkeletonConstants).VERSION}" />
+    </bean>
     <bean class="org.n52.iceland.skeleton.operator.GetCapabilitiesOperator" />
     <bean class="org.n52.iceland.skeleton.operator.DemoOperator" />
 
     <bean class="org.n52.iceland.skeleton.handler.impl.DemoHandlerImpl" />
+    <bean class="org.n52.iceland.skeleton.handler.impl.GetCapabilitiesHandlerImpl" />
 
 </beans>
